create folder for access integration tests & move slack notification access tests to separate file
